Rename scripts to not contain dashes.
