+++ scripts/tools/buildckeditor.js
