+++ scripts/tools/updateconfigfile.js
